refactor(compete-result): 승패 계산 로직을 CompeteResult가 담당하도록 수정
