+++ src/main/java/blackjack/domain/participant/Dealer.java
@@ -2,14 +2,9 @@ package blackjack.domain.participant;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.CardHand;
-import blackjack.domain.card.CompeteResult;
 import blackjack.domain.card.Deck;
 
 import java.util.ArrayList;
-import java.util.EnumMap;
-import java.util.List;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 public class Dealer extends Participant {
     
@@ -42,32 +37,5 @@ public class Dealer extends Participant {
     
     public Card getCard(int cardIndex) {
         return getCards().get(cardIndex);
-    }
-    
-    public EnumMap<CompeteResult, Long> competeResultMap(List<Player> players) {
-        return players.stream()
-                      .map(this::compete)
-                      .collect(Collectors.groupingBy(Function.identity(), () -> new EnumMap<>(CompeteResult.class),
-                              Collectors.counting()));
-    }
-    
-    private CompeteResult compete(Player player) {
-        if (isDealerWin(player)) {
-            return CompeteResult.WIN;
-        }
-        
-        if (isDealerDefeat(player)) {
-            return CompeteResult.DEFEAT;
-        }
-        
-        return CompeteResult.DRAW;
-    }
-    
-    private boolean isDealerWin(Player player) {
-        return (sumCardHand() > player.sumCardHand()) || player.isBurst();
-    }
-    
-    private boolean isDealerDefeat(Player player) {
-        return (sumCardHand() < player.sumCardHand()) && !player.isBurst();
     }
 }
